# api/swagger.yaml
openapi: 3.0.0
info:
  title: Rule Based Approval Engine API
  description: A comprehensive REST API for managing leave, expense, and discount approvals with rule-based evaluation
  version: 1.0.0
  contact:
    name: API Support
    email: support@approvalengine.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.approvalengine.com
    description: Production Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@company.com"
        grade_id:
          type: integer
          format: int64
          example: 2
        role:
          type: string
          enum: [employee, manager, admin]
          example: "employee"
        manager_id:
          type: integer
          format: int64
          nullable: true
          example: 5
        created_at:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Jane Smith"
        email:
          type: string
          format: email
          example: "jane@company.com"
        password:
          type: string
          minLength: 6
          example: "SecurePassword123"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "john@company.com"
        password:
          type: string
          example: "SecurePassword123"
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        message:
          type: string
          example: "Please provide valid input"
        code:
          type: integer
          example: 400
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "token=eyJhbGci...; HttpOnly; Secure; SameSite=Strict"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...